SQL statement composers ready for tests
DROP and CREATE table composers:
Now the table is DROPPED and recreated every time the flow is run.
The case where the tables are dropped, even if the data pull from the API
fails, is currently not handled! A new issue has been created for handling this:
https://github.com/Randers-Kommune-Digitalisering/jobcenterindblik_jobindsats.dk/issues/5

INSERT statment composer:
Moved the data to a new object 'data' and created a JSONata expression that:
1. Defines a set of helper functions to parenthesize strings and to add MariaDB
compatible quotes around strings (and skips numbers)
2. Processes the data and prepares it as valid SQL statmenent data strings,
using the helper functions
3. Composes a valid SQL 'INSERT INTO'statement string, from the prepared data strings